Transfered the Sorting functionalities to the question page
