+++ frontend/src/pages/QuestionPage.js
@@ -6,6 +6,8 @@ import {
 	AccordionSummary,
 	AccordionDetails,
 	IconButton,
+	MenuItem,
+	Select,
 } from '@mui/material';
 import { useParams } from 'react-router-dom';
 import Question from '../components/questions/Question';
@@ -26,6 +28,7 @@ import ReportDialog from '../components/dialogs/ReportDialog';
 import { useSnackbar } from 'notistack';
 import { ReactComponent as MessageReport } from '../assets/messageReport.svg';
 import isSwear from '../utils/isSwear';
+import globalTheme from '../globalTheme';
 const iconSelector = {
 	video: <VideoLibraryIcon />,
 	paper: <ArticleIcon />,
@@ -65,6 +68,35 @@ const useStyles = makeStyles(() => ({
 		padding: '1rem',
 	},
 }));
+function descendingComparator(a, b, orderBy) {
+	// We assume that values are descending
+	if (Number(b[orderBy]) < Number(a[orderBy])) {
+		// B<A logo A vai aparecer primeiro que B
+		return -1;
+	}
+	if (Number(b[orderBy]) > Number(a[orderBy])) {
+		// B>A logo B vai aparecer primeiro que A
+		return 1;
+	}
+	return 0;
+}
+
+// This method is created for cross-browser compatibility, if you don't
+// need to support IE11, you can use Array.prototype.sort() directly
+function stableSort(array, comparator) {
+	const stabilizedThis = array.map((el, index) => [el, index]); //Stores each object and it's index prior to sorting
+	stabilizedThis.sort((a, b) => {
+		//Sorts according to a comparator Function (if it returns value > 0 sort A after B) (if it returns value < 0 sort B after A)
+		//(If they are the same, the old indexes decide, and the same sorting is mantained for that case)
+		const order = comparator(a[0], b[0]);
+		if (order !== 0) {
+			return order;
+		}
+		return a[1] - b[1];
+	});
+	return stabilizedThis.map((el) => el[0]);
+	//Eliminates old index, makes it only one element again, only store the first element of the array which is the initial object
+}
 
 export default function QuestionPage() {
 	const classes = useStyles();
@@ -74,8 +106,26 @@ export default function QuestionPage() {
 	const [userID, setUserID] = useState(null);
 	const [isUserMod, setIsUserMod] = useState(false);
 	const [open, setOpen] = useState(false);
+	const [sortingType, setSortingType] = useState('votes');
+	const [comments, setComments] = useState([]);
 
 	const windowArray = useWindowDimensions();
+
+	const changeSorting = () => {
+		if (sortingType === 'votes') {
+			setSortingType('id');
+		} else {
+			setSortingType('votes');
+		}
+	};
+
+	const sortRerender = (newComments = comments) => {
+		setComments(
+			stableSort(newComments, (a, b) => descendingComparator(a, b, sortingType)).map(
+				(comment) => comment
+			)
+		);
+	};
 
 	const handleClickOpen = () => {
 		setOpen(true);
@@ -134,10 +184,16 @@ export default function QuestionPage() {
 		}
 	};
 	useEffect(() => {
-		fetchQuestion(id, (res) => {
-			setQuestion(res.data);
-		});
-	}, []);
+		if (question) {
+			sortRerender(comments);
+		} else {
+			console.log('hi there');
+			fetchQuestion(id, (res) => {
+				setQuestion(res.data);
+				sortRerender(res.data.comment);
+			});
+		}
+	}, [sortingType]);
 
 	useEffect(() => {
 		getUser((res1) => {
@@ -369,12 +425,61 @@ export default function QuestionPage() {
 			<Grid item xs={12}>
 				<Box>
 					<Grid container direction='column'>
+						<Grid
+							container
+							direction='row'
+							justifyContent='flex-end'
+							alignItems='center'
+						>
+							<Grid item>
+								<Select
+									labelId='sorter'
+									id='sorter'
+									value={sortingType}
+									onChange={changeSorting}
+									sx={globalTheme.components.select.styleOverrides}
+									MenuProps={{
+										sx: globalTheme.components.menuItem.styleOverrides,
+									}}
+								>
+									<MenuItem value={'votes'}>
+										<Typography
+											style={{
+												fontSize: responsiveWidth(
+													windowArray,
+													10,
+													20,
+													0.01
+												),
+											}}
+										>
+											Ordernar por Pontuação
+										</Typography>
+									</MenuItem>
+									<MenuItem value={'id'}>
+										<Typography
+											style={{
+												fontSize: responsiveWidth(
+													windowArray,
+													10,
+													20,
+													0.01
+												),
+											}}
+										>
+											Ordernar por Data
+										</Typography>
+									</MenuItem>
+								</Select>
+							</Grid>
+						</Grid>
 						{/* List of Comments Area */}
 						<Chat
 							userID={userID}
 							isUserMod={isUserMod}
 							questionID={id}
-							messageArray={question.comment}
+							messageArray={comments}
+							sortRerender={sortRerender}
 						></Chat>
 					</Grid>
 				</Box>
